Added Functionality for admin

+++ UserMicroservice/src/main/java/com/example/repo/UserTableRepo.java
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.example.entities.User;
 
 @Repository
-public interface UserTableRepo extends CrudRepository<User, Integer>  {
+public interface UserTableRepo extends CrudRepository<User, Long> {
     User findByUsernameAndPassword(String username, String password);
+
     User findByUsernameAndNickName(String username, String nickName);
 }
